requirements.txt
cycler==0.10.0
greenlet==0.4.17
kiwisolver==1.3.1
matplotlib==3.3.3
msgpack==1.0.0
numpy==1.19.4
Pillow==8.0.1
pynvim==0.4.2
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
